Remove erroneous long double casts from audiofile.cpp

+++ tools/audiofile/audiofile.cpp
@@ -5075,7 +5075,7 @@ bool ModuleState::fileModuleHandlesSeeking() const
 
 status ModuleState::setup(AFfilehandle file, Track *track)
 {
-	AFframecount fframepos = std::llrint((long double)track->nextvframe * track->f.sampleRate / track->v.sampleRate);
+	AFframecount fframepos = std::llrint(track->nextvframe * track->f.sampleRate / track->v.sampleRate);
 	bool isReading = file->m_access == _AF_READ_ACCESS;
 
 	if (!track->v.isUncompressed())
@@ -5146,11 +5146,11 @@ status ModuleState::setup(AFfilehandle file, Track *track)
 		if (track->totalfframes == -1)
 			track->totalvframes = -1;
 		else
-			track->totalvframes = std::llrint((long double)track->totalfframes *
+			track->totalvframes = std::llrint(track->totalfframes *
 				(track->v.sampleRate / track->f.sampleRate));
 
 		track->nextfframe = fframepos;
-		track->nextvframe = std::llrint((long double)fframepos * track->v.sampleRate / track->f.sampleRate);
+		track->nextvframe = std::llrint(fframepos * track->v.sampleRate / track->f.sampleRate);
 
 		m_isDirty = false;
